% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median-fast.R
\name{median_fast}
\alias{median_fast}
\title{Faster median function}
\usage{
median_fast(x)
}
\description{
This is a slimmed-down version of \code{\link[=median]{median()}} that doesn't do any checks and
only works on the data type we need.
}
\examples{
x <- runif(1000)
bench::mark(median(x), median_fast(x))
}
