% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median-value.R
\name{median_value}
\alias{median_value}
\title{The median of a/b of each node where a/b is between the bounds}
\usage{
median_value(tree, bounds, dfs_order)
}
\arguments{
\item{`tree`}{(\code{environment}) The whole tree, originally created by
\code{\link[=tree_from_edges]{tree_from_edges()}}.}

\item{`bounds`}{A list of \code{low} and \code{high} numeric values between which the
true maximum value of the tree is sought.  The bounds will initially be
\code{list(low = -Inf, high = Inf)}.}

\item{`dfs_order`}{(\code{character}) A vector of node IDs in depth-first search
order, from a \code{node$subtree_dfs$dfs}, created by \code{\link[=set_subtrees]{set_subtrees()}}.}
}
\value{
A numeric value that is the median of the values of all nodes whose value is
between \code{bounds$low} and \code{bounds$high} (inclusive), to be used by \code{\link[=value]{value()}}.
}
\description{
A numeric value that is the median of the values of all nodes whose value is
between \code{bounds$low} and \code{bounds$high} (inclusive).  This is used as an
estimate of the true maximum possible value of the tree.  The bounds are the
known limits of the true maximum.  The value of a single node is \code{a / b} of
its own values \code{a} and \code{b}.
}
