% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-child-into-parent.R
\name{merge_child_into_parent}
\alias{merge_child_into_parent}
\title{Merge a child node into its parent}
\usage{
merge_child_into_parent(child, dfs_env)
}
\arguments{
\item{`node`}{(\code{environment}) The node to be subsumed by its parent.}

\item{`dfs_env`}{(\code{environment}) An environment containing an object called
\code{dfs} that is a named list of \code{NULL} values, where the names are IDs of
nodes in depth-first order.  See \code{\link[=set_subtrees]{set_subtrees()}}.}
}
\value{
There is no return value. The tree is modified in place.
}
\description{
Nodes that are only-children may be merged into their parents, because
the matter of whether they are in the optimum tree depends entirely on
whether their parents are.
}
\details{
\enumerate{
\item Their values \code{a} and \code{b} are added to the parent's values \code{a} and \code{b}.
\item They are removed from the parent's \code{children} environment.
\item Their children, if any, are added to the parent's \code{children}
environment (their grandparent).
\item They are removed from the depth-first search order.
\item They are given a value \code{merged <- TRUE} for use by \code{\link[=propagate_value]{propagate_value()}},
which applies the averaged value of the eventual ancestor to the node.
}
}
