% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagate-value.R
\name{propagate_value}
\alias{propagate_value}
\title{Propagate values from parents to children (mutates the tree)}
\usage{
propagate_value(tree, dfs_order)
}
\arguments{
\item{`tree`}{(\code{environment}) Originally created by \code{\link[=tree_from_edges]{tree_from_edges()}}.}

\item{`dfs_order`}{(\code{character}) A vector of node IDs in depth-first search
order.}
}
\value{
There is no return value. Nodes are modified in place.
}
\description{
The value of a node is the the values of its ancestor, with which it has been
averaged.  This function propagates the values down the tree from ancestors
to descendants.
}
