% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-tree.R
\name{report_tree}
\alias{report_tree}
\title{Report the final state of the tree}
\usage{
report_tree(edges, tree)
}
\arguments{
\item{`edges`}{(\verb{data frame}) Originally created by \code{\link[=directed]{directed()}}.}

\item{`tree`}{(\code{environment}) Originally created by \code{\link[=tree_from_edges]{tree_from_edges()}}.}
}
\value{
A data frame, one row per node.  The same columns as \code{edges} (see
\code{\link[=directed]{directed()}}), plus:
\itemize{
\item \code{depth} The depth of the node in the original tree, before optimisation.
\item \code{ancestor} Of all the nodes that this node's value is averaged with, the ID
of ancestor that is closest to the root.  All the nodes between this node
and that ancestor have the same average value, and there might be other
descendants of that ancestor that do too.
\item \code{value} The average value of all the descendants of the \code{ancestor} that are
averaged together by \code{sum(a) / sum(b)}.
}
}
\description{
Takes the directed edges from \code{\link[=directed]{directed()}}, and the maximised \code{tree}, and
appends the tree values to the edges. The columns \code{from} and \code{to} are
converted back to \code{numeric}, which is how they would have been when they were
given to \code{weighted_maximum_mean_subtrees()}.
}
