% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore-subtree.R
\name{restore_subtree}
\alias{restore_subtree}
\title{Restore a node to its original state}
\usage{
restore_subtree(tree, dfs_order)
}
\arguments{
\item{`tree`}{(\code{environment}) Originally created by \code{\link[=tree_from_edges]{tree_from_edges()}}.}

\item{`dfs_order`}{(\code{character}) A vector of node IDs in depth-first search
order, from a \code{node$subtree_dfs$dfs}, created by \code{\link[=set_subtrees]{set_subtrees()}}.}
}
\value{
No value is returned.  The node is modified in place.
}
\description{
When the algorithm returns to a subtree that was pruned, it must return the
node's \code{parent}, \code{children} and values \code{a} and \code{b} to their original state.
This is done by copying \verb{original_*} values.  It is also done before the
first pass of the algorithm, because nodes are created with only \verb{original_*}
values.
}
