% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-subtrees.R
\name{set_subtrees}
\alias{set_subtrees}
\title{Give each node a list of its descendants}
\usage{
set_subtrees(tree, dfs_order)
}
\arguments{
\item{`tree`}{(\code{environment}) Originally from \code{\link[=directed]{directed()}}.}

\item{`dfs_order`}{(\code{character}) A vector of node IDs in depth-first search
order, originally from \code{\link[=directed]{directed()}}.}
}
\value{
No value is returned.  The nodes are modified in place.
}
\description{
After finding the optimal subtree connected with the root, some subtrees will
probably have been pruned.  These are then optimised in their own right,
recursively, so that the value of adding the to the optimum subtree is known.
To recurse into subtrees, each node keeps a list of its descendants in
depth-first search order, so that a depth-first search search doesn't have to
be performed for every pruned subtree.  The list is wrapped in an
environment, so that when it is passed to functions, the functions can modify
it in place.
}
