% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-from-edges.R
\name{tree_from_edges}
\alias{tree_from_edges}
\title{Create a tree of environments from a dataframe of edges}
\usage{
tree_from_edges(edges)
}
\arguments{
\item{`edges`}{A data frame with the columns \code{from} (ID of a node) and \code{to}
(ID of a child node of \code{from}), such as you get from
\code{\link[emstreeR:ComputeMST]{emstreeR::ComputeMST()}}.}
}
\value{
An environment containing one environment (node) per node ID in \code{edges},
either the \code{to} or the \code{from} column.
\itemize{
\item \code{id} (\code{character}).
\item \code{original_children}, an environment that is a list of environments, one for
each child of the node.  It is called \code{original_children} because copies
will later be made that might be pruned.
\item \code{merged}, a logical value initialised to \code{FALSE}.
}
}
\description{
Create a tree from a data frame, one row per edge.
}
\details{
Create a tree from a data frame, one row per edge, with the columns:
\itemize{
\item \code{from} ID of a node.
\item \code{to} ID of a child of \code{from}.
}

The edges don't have to be of a directed tree.  The tree will be made
directed later by \code{\link[=directed]{directed()}}.

For each row, an environment is created that contains the \code{id} of that row.
The environment also contains another environment, \code{children}, which contains
an environment for each child node.  Finally it contains a logical value
\code{merged}, initialised to \code{FALSE}.

One more environment is created to contain all the other environments
("nodes"), named by their \code{id}.  That environment is what this function
returns.
}
