% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-bounds.R
\name{update_bounds}
\alias{update_bounds}
\title{Narrow the bounds by comparing the tree's value with an estimated maximum}
\usage{
update_bounds(bounds, estimate, value)
}
\arguments{
\item{`bounds`}{A list of \code{low} and \code{high} numeric values between which the
true maximum value of the tree is sought.  The bounds will initially be
\code{list(low = -Inf, high = Inf)}.}

\item{`estimate`}{(\code{numeric}) An estimate of the maximum possible value of the
tree, from \code{\link[=median_value]{median_value()}}.}

\item{`value`}{(\code{numeric}) The value of the tree when greedily pruned by
removing nodes with values below \code{estimate}, from \code{\link[=value]{value()}}.}
}
\value{
An updated copy of \code{bounds}, to be used by \code{\link[=prune]{prune()}}.
}
\description{
The true maximum value of the tree is sought between lower and upper bounds.
}
\details{
The true maximum value of the tree is sought between lower and upper bounds.
The bounds can be narrowed by first estimating the true maximum with
\code{\link[=median_value]{median_value()}}, then greedily (but temporarily) pruning the tree to remove
nodes with values below that estimate, and then comparing the value of that
(temporarily) pruned tree with the estimate.  The pruning and valuing is done
in a single call to \code{\link[=value]{value()}}.

Depending on the \code{value} relative to the \code{estimate}:
\itemize{
\item The lower bound can always be raised to the value of the (temporarily) pruned
tree.
\item If the value of the (temporarily) pruned tree is equal to or below the
estimate, then the upper bound is lowered to the estimate, because it can be
proved that the estimate is an upper bound for the true maximum.
}
}
