% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{value}
\alias{value}
\title{The value of sum(a)/sum(b) of a tree omitting nodes below a cutoff}
\usage{
value(tree, cutoff, dfs_order)
}
\arguments{
\item{`tree`}{(\code{environment}) The whole tree, originally created by
\code{\link[=tree_from_edges]{tree_from_edges()}}.}

\item{`cutoff`}{(\code{numeric}) An estimate of the true maximum possible value of
the tree, used to greedily discard nodes.  Obtained from \code{\link[=median_value]{median_value()}}.}

\item{`dfs_order`}{(\code{character}) A vector of node IDs in depth-first search
order, from a \code{node$subtree_dfs$dfs}, created by \code{\link[=set_subtrees]{set_subtrees()}}.}
}
\value{
\itemize{
\item \code{value} (\code{numeric}) The value of the tree if it were pruned greedily by
this algorithm, to be used by \code{\link[=update_bounds]{update_bounds()}}.
}
}
\description{
This is a greedy algorithm to obtain a tree with a high average value, which
is \code{sum(a) / sum(b)} of all the nodes in the tree that remain after
temporarily discarding low-value nodes.  The tree is not modified in this
process.
}
\details{
Nodes are traversed in post-order (leaves first).
\itemize{
\item Nodes whose value \code{a / b} is below the cutoff are discarded by not adding
their values \code{a} and \code{b} to their parents' values \code{a} and \code{b}.
\item Nodes whose value \code{a / b} is above the cutoff add their value \code{a} to their
parent's value \code{a}, and their value \code{b} to their parents' value \code{b}.
}

Hence by the time the root node is reached, its \code{a} is now effectively
\code{sum(a)} of the undiscarded descendants, and similarly its \code{b} is \code{sum(b)} of
the undiscarded descendants, so its final value \code{a / b} is effectively
\verb{sum(a) / sum(b})`.
}
