% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmms.R
\name{weighted_maximum_mean_subtrees}
\alias{weighted_maximum_mean_subtrees}
\title{Find the highest possible sum(a)/sum(b) of any subtree}
\usage{
weighted_maximum_mean_subtrees(edges, root_id)
}
\arguments{
\item{`edges`}{A data frame of edges of the tree, with at least the columns
\code{to} and \code{from} (integer IDs of nodes), and \code{a} and \code{b} (numeric).  The
\code{from} and \code{to} IDs will be swapped where necessary to ensure that \code{to} is
the one that is further from the root node.  Then the edge values \code{a} and
\code{b} will be treated as belonging to the node rather than to the edge.
Passed to \code{\link[=tree_from_edges]{tree_from_edges()}}.}

\item{`root_id`}{(\code{numeric}) The ID of the root node of the tree.  Passed to
\code{\link[=directed]{directed()}} and other functions.}
}
\value{
The original \code{edges} data frame, with the \code{from} and \code{to} values swapped
where necessary to ensure the edges are directed away from the root node, and
with the additional columns:
\itemize{
\item \code{ancestor} Of all the nodes that a node's value is averaged with, the ID of
ancestor that is closest to the root.  All the nodes between this edge and
that ancestor have the same average value, and there might be other
descendants of that ancestor that do too.
\item \code{value} The average value of all the descendants of the \code{ancestor} that are
averaged together by \code{sum(a) / sum(b)}.  An nodes's \code{value} is less than
\code{average} if the node isn't part of the maximum subtree. But it is still
the highest possible for that node if it were added to the maximum subtree
with any others in its set.

Because there is still exactly one node per edge, the node values can be
interpreted as edge values again.
}
}
\description{
This is an exact algorithm to obtain a subtree, connected with the root, that
has the highest possible average value per node, which is \code{sum(a) / sum(b)}
of all the edges in the tree that remain after discarding low-value nodes.
}
\details{
For details of the algorithm, see the documentation for \code{wmms}.
}
